New implementation of the app for simplyfication
